v1.5.0: Update fwew-lib to 5.16, add /update route
Provides a path for fwedit or other fwew-api clients to interact with
the API for reloading the API cache from the latest fwedit MySQL data
(using env variables). This update is mostly for server-side
shenanigans. Anyone who would be enough of a nerd to unironically be
running fwew-api on their machine will not notice a difference from
before due to falling back on dictionary-v2.txt by default.

I hope that no one will abuse the /update route, because it causes a 3
second series of db reads and cache loading operations per call. Will
consider locking down or otherwise rate-limiting the /update endpoint if
it ends up being spammed in the future.